// src/app/page/master/pegawai-sk-loan/pegawai-sk-loan.interface.ts
import { EventEmitter } from '@angular/core';

export interface PegawaiSkLoan {
    noSK?: any;
    kdKategoriPegawai?: number;
    kdMasaKerja?: string;
    kdJabatanStruktural?: string;
    kdProduk?: string;
    kdKondisiProduk?: string;
    Resume?: string;
    Keterangan?: string;
    statusEnabled?: boolean;
    noRec?: string;
    label?: any;
    value?: any;
    namaStatus?: string;
    id?: any;
}